% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{adjust_N_samples}
\alias{adjust_N_samples}
\title{adjust N_samples for predictions}
\usage{
adjust_N_samples(N_samples, chain_length)
}
\arguments{
\item{N_samples}{original value for N_samples}

\item{chain_length}{length of NUTS chain}
}
\value{
integer: chain_length if N_samples == 0, min(N_samples, chain_length) otherwise
}
\description{
adjust N_samples for predictions
}
