% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{apply_named_args}
\alias{apply_named_args}
\title{a version of apply. if MARGIN = 1, the values in each column of X are passed to FUN
as named arguments according to the column name. if MARGIN = 2, the values in
each row of X are passed to FUN
as named arguments according to the row name}
\usage{
apply_named_args(X, MARGIN = 1, FUN, keep_class = TRUE)
}
\arguments{
\item{X}{see arguments for apply.  If MARGIN = 1, colnames(X) must match the named
arguments of FUN.}

\item{MARGIN}{see arguments for apply}

\item{FUN}{see arguments for apply.  Must have named arguments}
}
\value{
see apply
}
\description{
a version of apply. if MARGIN = 1, the values in each column of X are passed to FUN
as named arguments according to the column name. if MARGIN = 2, the values in
each row of X are passed to FUN
as named arguments according to the row name
}
