% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_burst_size_homogeneous}
\alias{calc_burst_size_homogeneous}
\title{calculate burst size for homogeneous model}
\usage{
calc_burst_size_homogeneous(pars)
}
\arguments{
\item{pars}{named vector or list of parameter values with elements "beta1", "T_0", "p", "delta", "c1"
if there is heterogeneity in the model, pars needs to be a list, and beta1, T_0, p can be vectors}
}
\value{
scalar: value of r
}
\description{
calculate burst size for homogeneous model
}
