% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calc_mean_beta}
\alias{calc_mean_beta}
\title{calculate inferred mean value of beta}
\usage{
calc_mean_beta(betas)
}
\arguments{
\item{betas}{tibble with columns iter, par_name, value, chain, ... where par_name is beta1\link{1} or beta1\link{2}}
}
\value{
tibble with columns iter, chain, ..., mean_beta
}
\description{
calculate inferred mean value of beta
}
