% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calc_prop}
\alias{calc_prop}
\title{calculate proportion of each cell type and infection status for each time point and replicate}
\usage{
calc_prop(sol)
}
\arguments{
\item{sol}{tibble with columns t, sigma, replicate, compartment_name and value}
}
\value{
tibble with columns t, sigma, replicate, compartment_name and value
}
\description{
calculate proportion of each cell type and infection status for each time point and replicate
}
