% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_r_homogeneous}
\alias{calc_r_homogeneous}
\title{calculate r for homogeneous case (faster)}
\usage{
calc_r_homogeneous(beta1, T_0, p, delta, c1)
}
\arguments{
\item{beta1}{value of beta1}

\item{T_0}{value of T_0}

\item{p}{value of p}

\item{delta}{value of delta}

\item{c1}{value of c1}
}
\value{
scalar: value of R_0
}
\description{
calculate r for homogeneous case (faster)
}
