% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{calc_r_teiv}
\alias{calc_r_teiv}
\title{fit model to data using cmdstan}
\usage{
calc_r_teiv(beta_T_0, p, k, delta, c1)
}
\arguments{
\item{beta_T_0}{scalar}

\item{p}{scalar}

\item{k}{scalar}

\item{delta}{scalar}

\item{c1}{scalar}
}
\description{
fit model to data using cmdstan
}
