% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calc_weighted_mean_beta}
\alias{calc_weighted_mean_beta}
\title{calculate inferred weighted mean value of beta (weighted by $T_0i$)}
\usage{
calc_weighted_mean_beta(betas, samples)
}
\arguments{
\item{betas}{tibble with columns iter, par_name, value, chain, ... where par_name is beta1\link{1} or beta1\link{2}}

\item{samples}{tibble with columns alpha, ... where ... are the same indexing variables as in betas}
}
\value{
tibble with columns iter, chain, weighted_mean_beta, ...
}
\description{
calculate inferred weighted mean value of beta (weighted by $T_0i$)
}
