% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{change_sigma_wrapper}
\alias{change_sigma_wrapper}
\title{solve TIV model with heterogeneous betas for different values of sigma}
\usage{
change_sigma_wrapper(solve_TIV_two_beta, pars, sigma)
}
\arguments{
\item{solve_TIV_two_beta}{function to solve TIV model with heterogeneous betas}

\item{pars}{list of parameter values from get_two_compartment_pars}

\item{sigma}{scalar}
}
\value{
tibble with model solution
}
\description{
solve TIV model with heterogeneous betas for different values of sigma
}
