% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{enforce_beta_larger}
\alias{enforce_beta_larger}
\title{take tibble of inferred beta values and enforce beta1 > beta2}
\usage{
enforce_beta_larger(betas)
}
\arguments{
\item{betas}{tibble with columns iter, par_name, value, chain, sigma where par_name is beta1\link{1} or beta1\link{2}}
}
\value{
tibble with columns iter, par_name, value, chain, sigma
}
\description{
take tibble of inferred beta values and enforce beta1 > beta2
}
