% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{extract_fit}
\alias{extract_fit}
\title{extract samples from fit object and thin}
\usage{
extract_fit(fit, N_samples, par_names, drop = FALSE, by_chain = FALSE)
}
\arguments{
\item{fit}{stanfit object}

\item{N_samples}{number of samples to extract.  if 0 or greater than the number of available samples, extract all}

\item{par_names}{(optional) character vector: namse of parameters to extract.  if missing, extract all}

\item{drop}{if par_names is specified and is of length 1, and drop = TRUE, return vector;
otherwise return matrix}

\item{by_chain}{if TRUE, return list of matrices with one matrix per chain.  Then N_samples is interpreted as per
chain.  if FALSE, return one matrix for the whole chain.}
}
\value{
matrix with named columns containing samples
}
\description{
extract samples from fit object and thin
}
