% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R, R/summary_stats.R
\name{extract_par_values}
\alias{extract_par_values}
\title{new function for extracting parameter values (to do: compare with old)}
\usage{
extract_par_values(fit, par_names)

extract_par_values(fit, par_names)
}
\arguments{
\item{fit}{fit object from RStan or cmdstanr}

\item{par_name}{parameter names in fit}
}
\value{
tibble with columns iter, chain, value, par_name

tibble with columns iter, chain, value, par_name
}
\description{
new function for extracting parameter values (to do: compare with old)

new function for extracting parameter values (to do: compare with old)
}
