% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gen_prior_indiv_beta}
\alias{gen_prior_indiv_beta}
\title{sample from prior distribution for the ratio of individual beta to beta}
\usage{
gen_prior_indiv_beta(n_C = 2, n_samples_prior = 10000)
}
\arguments{
\item{n_C}{number of cell types}
}
\value{
tibble with columns par_name, value and sigma = "prior"
}
\description{
sample from prior distribution for the ratio of individual beta to beta
}
