% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gen_prior_mean_beta}
\alias{gen_prior_mean_beta}
\title{sample from prior distribution for the mean of beta}
\usage{
gen_prior_mean_beta(n_C = 2, n_samples_prior = 10000)
}
\arguments{
\item{n_C}{number of cell types}
}
\value{
tibble with columns mean_beta (contains the samples)
}
\description{
sample from prior distribution for the mean of beta
}
