% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_betas}
\alias{get_betas}
\title{read in inferred values of beta}
\usage{
get_betas(filename_fn, ...)
}
\arguments{
\item{filename_fn}{function which
returns the filename of the fit}

\item{...}{arguments to be passed to filename_fn}
}
\value{
tibble with columns iter, par_name, value, chain, sigma where par_name is beta1\link{1} or beta1\link{2}
}
\description{
read in inferred values of beta
}
