% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars.R
\name{get_git_dir}
\alias{get_git_dir}
\title{get directory of git repository}
\usage{
get_git_dir(spartan)
}
\arguments{
\item{spartan}{logical: running on cluster or not}
}
\value{
directory name
}
\description{
get directory of git repository
}
