% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R, R/summary_stats.R
\name{get_max_LL_pars}
\alias{get_max_LL_pars}
\title{get the maximum likelihood parameters from a stan fit}
\usage{
get_max_LL_pars(fit_obj)

get_max_LL_pars(fit_obj)
}
\arguments{
\item{fit_obj}{a stanfit object, matrix containing draws from posterior, or output from cmdstanr}
}
\value{
named vector with parameters and generated quantities of maximum likelihood

named vector with parameters and generated quantities of maximum likelihood
}
\description{
get the maximum likelihood parameters from a stan fit

get the maximum likelihood parameters from a stan fit
}
