% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R, R/summary_stats.R
\name{get_model_filename}
\alias{get_model_filename}
\title{get model filename}
\usage{
get_model_filename(model_name)

get_model_filename(model_name)
}
\arguments{
\item{model_name}{string}
}
\value{
string: model filename

string: model filename
}
\description{
get model filename

get model filename
}
