% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{get_n_iter}
\alias{get_n_iter}
\title{get the number of iterations post warm-up for each chain in a stanfit object}
\usage{
get_n_iter(fit)
}
\arguments{
\item{fit}{stanfit object}
}
\value{
vector: iterations post warm-up for each chain
}
\description{
get the number of iterations post warm-up for each chain in a stanfit object
}
