% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_pars}
\alias{get_pars}
\title{read in inferred values of parameters}
\usage{
get_pars(filename_fn, par_names, ...)
}
\arguments{
\item{filename_fn}{function which returns the filename of the fit}

\item{par_names}{parmaeter names}

\item{...}{arguments to be passed to filename_fn}
}
\value{
tibble with columns iter, par_name, value, chain, sigma where par_name is beta1\link{1} or beta1\link{2}
}
\description{
read in inferred values of parameters
}
