% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars.R
\name{get_sampling_pars}
\alias{get_sampling_pars}
\title{get parameter values for sampling}
\usage{
get_sampling_pars()
}
\value{
list of parameter values
}
\description{
get parameter values for sampling
}
