% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars.R
\name{get_sampling_times}
\alias{get_sampling_times}
\title{get times for sampling}
\usage{
get_sampling_times()
}
\value{
named list of sampling times for cells and viral load
}
\description{
get times for sampling
}
