% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_sol_V}
\alias{get_sol_V}
\title{get model predictions for viral load}
\usage{
get_sol_V(filename, pred_times, used_pred)
}
\arguments{
\item{filename}{filename}

\item{used_pred}{if TRUE, predictions were made after model fitting; if FALSE, predictions
were made alongside model fitting}

\item{dir_name}{directory name}

\item{data_list}{data list for stan fitting}
}
\value{
tibble with columns idx, condition, t, V
}
\description{
get model predictions for viral load
}
