% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{get_summaries_subset_chains}
\alias{get_summaries_subset_chains}
\title{#' make model predictions
#'
#' @param model stan model object
#' @param data_list list of data and model parameters to feed to model
#' @return stanfit object
#' @export
make_pred <- function(model, data_list) {
if(is.character(model)) {
model <- rstan::stan_model(get_model_filename(model))
}
pred <- rstan::sampling(model,
data = data_list,
chains = 1, iter = 1,
algorithm = "Fixed_param")
pred
}}
\usage{
get_summaries_subset_chains(idx, Delta_scaled, fit_wrapper)
}
\arguments{
\item{idx}{index of simulation estimation parameter set}

\item{Delta_scaled}{value of delta_scaled used for simulation}

\item{fit_wrapper}{function to fit model to data (in practice used to retrive filename)}
}
\value{
tibble with columns idx (from input); Delta_scaled (from input);
prop_divergent (proportion of divertent transitions in selected chains);
max_rhat (maximum rhat calculated using selected chains);
chosen_chain_combn (indices of chosen chains)
}
\description{
choose subset of 8 chains from MCMC so that there are no divergences and
Rhat is below 1.01 (if possible) or minimised
}
