% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_true_values_beta_tibble}
\alias{get_true_values_beta_tibble}
\title{make a tibble with the true values of beta1\link{1} and beta1\link{2}}
\usage{
get_true_values_beta_tibble(beta1, sigma_vec, n_C = 2)
}
\arguments{
\item{beta1}{scalar: value of beta1}

\item{sigma_vec:}{vector of values of sigma}
}
\value{
tibble with columns par_name, value and sigma
}
\description{
make a tibble with the true values of beta1\link{1} and beta1\link{2}
}
