% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars.R
\name{get_two_compartment_pars}
\alias{get_two_compartment_pars}
\title{get two compartment parameters}
\usage{
get_two_compartment_pars(pars)
}
\arguments{
\item{pars}{named vector of parameter values}
}
\value{
list of parameter values
}
\description{
get two compartment parameters
}
