% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{impose_observation_threshold}
\alias{impose_observation_threshold}
\title{function which imposes an observation threshold on data.}
\usage{
impose_observation_threshold(noisy_values, threshold)
}
\arguments{
\item{noisy_values}{numeric vector of noisy values on which to impose threshold}

\item{threshold}{below this threshold, set values to -1 to denote below threshold}
}
\value{
numeric vector of values with imposed threshold
}
\description{
\code{impose_observation_threshold} imposes an observation threshold on data.
}
\details{
Run this after imposing other noise (e.g. lognormal)
}
