% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R, R/summary_stats.R
\name{index_par_name}
\alias{index_par_name}
\title{append indices to parameter name}
\usage{
index_par_name(par_name, idx)

index_par_name(par_name, idx)
}
\arguments{
\item{par_name}{string: parameter name}

\item{idx}{numeric vector}
}
\value{
string.  e.g. if par_name == "abc" and idx == c(1,2), returns "abc\link{1,2}"

string.  e.g. if par_name == "abc" and idx == c(1,2), returns "abc\link{1,2}"
}
\description{
append indices to parameter name

append indices to parameter name
}
