% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{make_pred_beta}
\alias{make_pred_beta}
\title{make model predictions}
\usage{
make_pred_beta(filename_fn, make_pred_fn, sigma)
}
\arguments{
\item{filename_fn}{function to get filename where parameter values are stored.  Takes argument sigma.}

\item{make_pred_fn}{function to make model predictions. Has arguments idx, beta1 and beta2,
and returns a tibble with columns idx, t and compartment_name and value.}

\item{sigma}{scalar value of sigma.}
}
\value{
pred tibble containing model predictions.
contains columns t, lower (2.5\% quantile of prediction),
upper (97.5\% quantile of prediction), max_LL (maximum likelihood prediction), sigma
}
\description{
make model predictions
}
