% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{plot_beta_ratios}
\alias{plot_beta_ratios}
\title{plot ratio of individual beta to mean beta}
\usage{
plot_beta_ratios(betas, mean_betas, true_values)
}
\arguments{
\item{betas}{tibble with columns iter, par_name, value, chain, sigma where par_name is beta1\link{1} or beta1\link{2}}

\item{mean_betas}{tibble with columns iter, chain, sigma, mean_beta}

\item{true_value}{true values of mean beta}
}
\value{
plot
}
\description{
plot ratio of individual beta to mean beta
}
