% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R, R/summary_stats.R
\name{plot_hist_stan}
\alias{plot_hist_stan}
\title{new function for plotting histograms (to do: compare with old)}
\usage{
plot_hist_stan(
  fit,
  par_names,
  by_chain = TRUE,
  true_values,
  prior_min,
  prior_max
)

plot_hist_stan(
  fit,
  par_names,
  by_chain = TRUE,
  true_values,
  prior_min,
  prior_max
)
}
\arguments{
\item{fit}{fit object from RStan or cmdstanr}

\item{par_names}{vector of parameter names in fit to plot}

\item{by_chain}{logical. Fill histogram by chain or not.}

\item{true_values}{(optional) vector of true values.  Must be of same length as par_names}

\item{prior_min}{(optional) vector of x axis lower limits.  Must be of same length as par_names}

\item{prior_max}{(optional) vector of x axis upper limits.  Must be of same length as par_names}
}
\value{
ggplot object

ggplot object
}
\description{
new function for plotting histograms (to do: compare with old)

new function for plotting histograms (to do: compare with old)
}
