% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{plot_mean_betas_histogram}
\alias{plot_mean_betas_histogram}
\title{plot mean betas}
\usage{
plot_mean_betas_histogram(mean_betas, true_value)
}
\arguments{
\item{mean_betas}{tibble with columns iter, chain, sigma, mean_beta}

\item{true_value}{true_value of mean beta}
}
\value{
plot
}
\description{
plot mean betas
}
