% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{plot_pred_V}
\alias{plot_pred_V}
\title{plot model predictions for viral load}
\usage{
plot_pred_V(pred, all_sim_data, lod_V)
}
\arguments{
\item{pred}{tibble containing model predictions.
contains columns t, lower (2.5\% quantile of prediction),
upper (97.5\% quantile of prediction), max_LL (maximum likelihood prediction), sigma}

\item{all_sim_data}{tibble with simulated data.
contains columns t, obs_V, sigma}

\item{lod_V}{lower limit of detection for viral load}
}
\value{
ggplot object
}
\description{
plot model predictions for viral load
}
