% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{plot_single_sol}
\alias{plot_single_sol}
\title{plot viral load from data and model fits}
\usage{
plot_single_sol(sol_V, sim_data)
}
\arguments{
\item{sol_V}{model prediction for viral load}

\item{sim_data}{simulated data}
}
\value{
ggplot object
}
\description{
plot viral load from data and model fits
}
