% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{plot_trace_log_lik}
\alias{plot_trace_log_lik}
\title{make traceplot of the log likelihood}
\usage{
plot_trace_log_lik(filename_fn, sigma)
}
\arguments{
\item{filename_fn}{function to get filename where parameter values are stored.  Takes argument sigma.}

\item{sigma}{scalar value of sigma.}
}
\value{
ggplot object
}
\description{
make traceplot of the log likelihood
}
