% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{prob_round}
\alias{prob_round}
\title{probabilistic rounding}
\usage{
prob_round(x)
}
\arguments{
\item{x}{scalar}
}
\value{
x rounded up with probability x - floor(x)
}
\description{
probabilistic rounding
}
