% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rl10norm}
\alias{rl10norm}
\title{Analogue for rlnorm but in base 10}
\usage{
rl10norm(n, mean = 1, sd = 1)
}
\arguments{
\item{n}{a numeric vector of length 1: number of observations}

\item{mean}{a vector of means.}

\item{sd}{a vector of standard deviations.}
}
\value{
a numeric vector of length n, distributed according to the log10normal distribution.
The log10normal distribution in log10 space has mu =  mean and sigma = sd.
}
\description{
\code{rl10norm} is an analogue for rlnorm but in base 10
}
