% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_dirichlet_multinomial}
\alias{sim_dirichlet_multinomial}
\title{simulate dirichlet multinomial observations}
\usage{
sim_dirichlet_multinomial(true_n_cells_vec, rho, sample_size, n_sims = 1)
}
\arguments{
\item{true_n_cells_vec}{vector of length 2 * n_C, where n_C is the number of cell types.
concatenated vector of number of uninfected cells of each type and
number of infected cells of each type in the well}

\item{rho}{overdispersion parameter between 0 and 1}

\item{sample_size}{sample size}
}
\value{
n_sims x 2 * n_C matrix, where n_C is the number of cell types.
concatenated vector of number of uninfected cells of each type and
number of infected cells of each type in the well.
if n_sims == 1, returns vector instead
}
\description{
simulate dirichlet multinomial observations
}
