% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate data from model with heterogeneous beta}
\usage{
simulate_data(sol, sampling_times, sampling_pars, cells, sim_filename)
}
\arguments{
\item{sol}{tibble with columns t and V -- solution for viral load}

\item{sampling_times}{vector of sampling times -- must be in sol$t}

\item{sampling_pars}{list if sampling parameters including lod_V (limit of detection) and sigma_V (sd of lognormal noise)}

\item{cells}{logical: whether to simulate observations for cells}

\item{sim_filename}{file in which to store simulation}
}
\value{
tibble with model solution
}
\description{
simulate data from model with heterogeneous beta
}
